Rename era pack to e2store & add support for E2HS file format (#3954)
* reorganize package into directory per format

* redefine Version and BlockIndex as CommonTypes

* define e2hs types

* move blockIndex functions to shared blockIndex.ts file

* write createBlockIndex function and replace code in era1

* implement e2hs format

* rename parse to decompress

* write parse function

* fix up blockTuple functions

* update blockTuple calls in e2hs code

* log better error on type mismatch

* test using reference file

* fix imports in tests

* rename era package to e2store

---------
