+++ packages/e2store/.c8rc.json
@@ -1,0 +1,4 @@
+{
+  "extends": "../../config/.c8rc.json",
+  "include": ["src/**/*.ts"]
+}
