+++ packages/e2store/src/e2hs/index.ts
@@ -1,0 +1,2 @@
+export * from './e2hs.ts'
+export * from './blockTuple.ts'
