+++ packages/e2store/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../config/tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./dist"
+  },
+  "include": ["src/**/*.ts", "test/**/*.ts"]
+}
