+++ packages/e2store/tsconfig.lint.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../../config/tsconfig.lint.json"
+}
