+++ packages/e2store/tsconfig.prod.cjs.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../config/tsconfig.prod.cjs.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "dist/cjs"
+  },
+  "include": ["src/**/*.ts"],
+  "references": [{ "path": "../rlp/tsconfig.prod.cjs.json" }]
+}
