+++ packages/e2store/tsconfig.prod.esm.json
@@ -1,0 +1,13 @@
+{
+  "extends": "../../config/tsconfig.prod.esm.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "dist/esm"
+  },
+  "include": ["src/**/*.ts"],
+  "references": [
+    { "path": "../rlp/tsconfig.prod.esm.json" },
+    { "path": "../block/tsconfig.prod.esm.json" },
+    { "path": "../blockchain/tsconfig.prod.esm.json" }
+  ]
+}
