+++ packages/e2store/typedoc.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: '../../config/typedoc.cjs',
+  entryPoints: ['src'],
+  out: 'docs',
+  exclude: ['test/*.ts'],
+}
