--- packages/era/.c8rc.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../../config/.c8rc.json",
-  "include": ["src/**/*.ts"]
-}
