--- packages/era/.gitignore
@@ -1,1 +1,0 @@
-
