--- packages/era/eslint.config.mjs
@@ -1,19 +1,0 @@
-import rootConfig from '../../config/eslint.config.mjs'
-
-export default [
-  ...rootConfig,
-  {
-    languageOptions: {
-      parserOptions: {
-        project: ['./tsconfig.lint.json'],
-      },
-    },
-  },
-  {
-    files: ['examples/**/*'],
-    rules: {
-      'no-console': 'off',
-      '@typescript-eslint/no-unused-vars': 'off',
-    },
-  },
-]
