--- packages/era/package.json
@@ -1,69 +1,0 @@
-{
-  "name": "@ethereumjs/era",
-  "version": "10.0.0-dev-rc.1",
-  "description": "Era file support for EthereumJS",
-  "keywords": ["ethereum", "era", "era1"],
-  "homepage": "https://github.com/ethereumjs/ethereumjs-monorepo/tree/master/packages/era#readme",
-  "bugs": {
-    "url": "https://github.com/ethereumjs/ethereumjs-monorepo/issues?q=is%3Aissue+label%3A%22package%3A+era%22"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ethereumjs/ethereumjs-monorepo.git"
-  },
-  "license": "MIT",
-  "author": "EthereumJS Team",
-  "contributors": [
-    {
-      "name": "ScottyPoi",
-      "url": "https://github.com/scottypoi"
-    }
-  ],
-  "type": "module",
-  "sideEffects": false,
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "exports": {
-    ".": {
-      "import": {
-        "typescript": "./src/index.ts",
-        "default": "./dist/esm/index.js"
-      },
-      "require": "./dist/cjs/index.js"
-    }
-  },
-  "files": ["dist", "src"],
-  "scripts": {
-    "biome": "npx @biomejs/biome check",
-    "biome:fix": "npx @biomejs/biome check --write",
-    "build": "../../config/cli/ts-build.sh",
-    "clean": "../../config/cli/clean-package.sh",
-    "coverage": "DEBUG=ethjs npx vitest run -c ../../config/vitest.config.coverage.mts",
-    "coverage:istanbul": "DEBUG=ethjs npx vitest run -c ../../config/vitest.config.coverage.istanbul.mts",
-    "docs:build": "npx typedoc --options typedoc.cjs",
-    "examples": "tsx ../../scripts/examples-runner.ts -- util",
-    "examples:build": "npx embedme README.md",
-    "lint": "npm run biome && eslint  --config ./eslint.config.mjs .",
-    "lint:fix": "npm run biome:fix && eslint  --fix --config ./eslint.config.mjs .",
-    "prepublishOnly": "../../config/cli/prepublish.sh",
-    "sc": "npm run spellcheck",
-    "spellcheck": "npm run spellcheck:ts && npm run spellcheck:md",
-    "spellcheck:ts": "npx cspell --gitignore -c ../../config/cspell-ts.json \"./**/*.ts\" --cache --show-suggestions --show-context",
-    "spellcheck:md": "npx cspell --gitignore -c ../../config/cspell-md.json \"**.md\" --cache --show-suggestions --show-context",
-    "test": "npm run test:node",
-    "test:node": "npx vitest run -c ../../config/vitest.config.mts",
-    "tsc": "../../config/cli/ts-compile.sh"
-  },
-  "dependencies": {
-    "@chainsafe/snappy-stream": "^5.1.2",
-    "@ethereumjs/block": "^10.0.0-rc.1",
-    "@ethereumjs/blockchain": "^10.0.0-rc.1",
-    "@ethereumjs/rlp": "^10.0.0-rc.1",
-    "@ethereumjs/util": "^10.0.0-rc.1",
-    "level": "^9.0.0",
-    "micro-eth-signer": "^0.14.0"
-  },
-  "engines": {
-    "node": ">=18"
-  }
-}
