--- packages/era/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../config/tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./dist"
-  },
-  "include": ["src/**/*.ts", "test/**/*.ts"]
-}
