--- packages/era/tsconfig.lint.json
@@ -1,3 +1,0 @@
-{
-  "extends": "../../config/tsconfig.lint.json"
-}
