--- packages/era/typedoc.cjs
@@ -1,6 +1,0 @@
-module.exports = {
-  extends: '../../config/typedoc.cjs',
-  entryPoints: ['src'],
-  out: 'docs',
-  exclude: ['test/*.ts'],
-}
